labs/multimodal-server/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pillow>=10.3.0
pydantic>=2.6.0
numpy>=1.26.0
httpx>=0.27.0
